parametricboot: Parametric Bootstrap for Fitted Regression Models

Automates the parametric bootstrap for models fitted with stats::lm(), stats::glm(), lme4::lmer(), lme4::glmer() and survival::coxph().

Workflow

Fit a model as usual.

Draw bootstrap replicates with pb_simulate(), or pb_parallel() when a single refit is slow.

Summarise them: pb_summary_table(), pb_key_stats(), pb_confint() and pb_tidy() for tables; pb_plot_estimates() and pb_plot_diagnostics() for plots. If some refits misbehaved, see pb_drop_warned().

Propagate the uncertainty to predictions with pb_predict_boot() and pb_plot_predictions().

Test nested models with the bootstrap likelihood-ratio test pb_lrt(), and compare models that are not nested with pb_compare_models().

See also

Author

Maintainer: Diogo Ribeiro diogo.debastos.ribeiro@gmail.com (ORCID) (ESMAD, Instituto Politecnico do Porto) [copyright holder]